requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

torch == 1.11.0+cu113
numpy == 1.22.3
pandas == 1.4.0
matplotlib == 3.1.2
pyyaml == 6.0
wandb == 0.12.9
plotly == 5.6.0
protobuf==3.20.*